fix eslint warn, restructure folder src, add routing system --wip,
