+++ .eslintrc.js
@@ -6,7 +6,13 @@ module.exports = {
   },
   parserOptions: {
     ecmaFeatures: {
+      // * kalo kamu pake file dengan ekstensi (.jsx) maka tambahkan baris ini
       jsx: true,
+
+      // * kalo kamu pake file dengan ekstensi (.tsx) maka tambahkan baris ini
+      // tsx: true,
     },
   },
+  // fix eslint  Either disable config file checking with requireConfigFile: false, or configure Babel so that it can find the config files.
+  requireConfigFile: false,
 };
