+++ index.js
@@ -3,11 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './App';
 import {name as appName} from './app.json';
+import App from './src/App';
 
-// default punya kamu
-// AppRegistry.registerComponent(appName, () => App);
-
-// panggil file yg dari dosen
 AppRegistry.registerComponent(appName, () => App);
