+++ package.json
@@ -10,13 +10,16 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
-    "lottie-react-native": "^6.0.0-rc.3",
     "react": "18.2.0",
     "react-native": "0.71.3",
     "react-native-safe-area-context": "^4.5.0",
-    "react-native-screens": "^3.20.0"
+    "react-native-screens": "^3.20.0",
+    "react-native-svg": "^13.8.0",
+    "react-native-svg-transformer": "^1.0.0",
+    "react-native-vector-icons": "^9.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
