+++ src/App.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
-import Router from './src/router';
+import Router from './router';
 
 function App() {
   return (
