+++ src/screens/home/Home.jsx
@@ -1,0 +1,12 @@
+import {View, Text} from 'react-native';
+import React from 'react';
+
+const Home = () => {
+  return (
+    <View>
+      <Text>Home</Text>
+    </View>
+  );
+};
+
+export default Home;
